refactor: App NavHost 보조 정책 분리
- BottomBar item 구성과 선택 상태 판별을 별도 파일로 분리

- Favorites root 재진입 정책을 하단 탭 navigation 정책으로 이동

- 즐겨찾기 장소 MapRoute 변환 로직을 mapper 파일로 분리

- app compile과 app unit test로 navigation 정책 유지 확인

Refs #138
