+++ app/src/main/java/com/team/yeogibeoryeo/navigation/YeogiBeoryeoNavHost.kt
@@ -19,7 +19,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
@@ -28,19 +27,10 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
-import java.util.UUID
 import com.team.yeogibeoryeo.common.navigation.AppBottomNavigationBar
-import com.team.yeogibeoryeo.common.navigation.BottomNavigationItem
-import com.team.yeogibeoryeo.common.navigation.navigateBottomTab
 import com.team.yeogibeoryeo.presentation.favorites.FavoritesRoute as FavoritesScreenRoute
-import com.team.yeogibeoryeo.presentation.favorites.model.FavoriteCollectionSpotMapMoveRequest
 import com.team.yeogibeoryeo.presentation.map.CollectionSpotMapScreen
-import com.team.yeogibeoryeo.presentation.map.model.FavoriteSpotMapMoveRequest
 import com.team.yeogibeoryeo.presentation.regionalguide.RegionalGuideRoute as RegionalGuideScreenRoute
-import com.team.yeogibeoryeo.domain.spot.model.CollectionSpotType
-import com.team.yeogibeoryeo.domain.spot.model.Coordinate
-import com.team.yeogibeoryeo.common.R as CommonR
-import com.team.yeogibeoryeo.R as AppR
 import com.team.yeogibeoryeo.presentation.search.ItemGuideDetailRoute as ItemGuideDetailScreenRoute
 import com.team.yeogibeoryeo.presentation.search.ItemSearchRoute as ItemSearchScreenRoute
 
@@ -80,57 +70,11 @@ fun YeogiBeoryeoNavHost(
                 exit = exitTransition,
             ) {
                 AppBottomNavigationBar(
-                    items =
-                        listOf(
-                            BottomNavigationItem(
-                                label = "품목",
-                                iconResId = CommonR.drawable.ic_symbol_recycle,
-                                selected = currentBackStackEntry.isItemSearchSelected(),
-                                onClick = {
-                                    navController.navigateBottomTabClearingFavoriteReentry(
-                                        currentBackStackEntry = currentBackStackEntry,
-                                        route = ItemSearchRoute(),
-                                    )
-                                },
-                            ),
-                            BottomNavigationItem(
-                                label = "지도",
-                                iconResId = AppR.drawable.ic_navigation_map,
-                                selected = currentDestination?.hasRoute<MapRoute>() == true,
-                                onClick = {
-                                    isBottomBarVisible = true
-                                    navController.navigateBottomTabClearingFavoriteReentry(
-                                        currentBackStackEntry = currentBackStackEntry,
-                                        route = MapRoute(),
-                                    )
-                                },
-                            ),
-                            BottomNavigationItem(
-                                label = "안내",
-                                iconResId = AppR.drawable.ic_navigation_guide,
-                                selected = currentBackStackEntry.isRegionalGuideSelected(),
-                                onClick = {
-                                    navController.navigateBottomTabClearingFavoriteReentry(
-                                        currentBackStackEntry = currentBackStackEntry,
-                                        route = RegionalGuideRoute(),
-                                    )
-                                },
-                            ),
-                            BottomNavigationItem(
-                                label = "저장",
-                                iconResId = CommonR.drawable.ic_favorite,
-                                selected = currentBackStackEntry.isFavoritesSelected(),
-                                onClick = {
-                                  if (currentBackStackEntry.isFavoriteRegionalGuideSelected()) {
-                                      navController.popBackStack<FavoritesRoute>(inclusive = false)
-                                  } else {
-                                      navController.navigateFavoritesRoot(
-                                          currentBackStackEntry = currentBackStackEntry,
-                                      )
-                                  }
-                                },
-                            ),
-                        ),
+                    items = navController.createBottomNavigationItems(
+                        currentBackStackEntry = currentBackStackEntry,
+                        currentDestination = currentDestination,
+                        onMapTabSelected = { isBottomBarVisible = true },
+                    ),
                 )
             }
         },
@@ -217,99 +161,4 @@ fun YeogiBeoryeoNavHost(
             }
         }
     }
-}
-
-private fun NavBackStackEntry?.isItemSearchSelected(): Boolean =
-    this?.destination?.hasRoute<ItemSearchRoute>() == true ||
-        isItemGuideDetailSource(ItemGuideDetailSource.SEARCH)
-
-private fun NavBackStackEntry?.isFavoritesSelected(): Boolean =
-    this?.destination?.hasRoute<FavoritesRoute>() == true ||
-        isItemGuideDetailSource(ItemGuideDetailSource.FAVORITES) ||
-        isFavoriteRegionalGuideSelected()
-
-private fun NavBackStackEntry?.isRegionalGuideSelected(): Boolean =
-    this?.destination?.hasRoute<RegionalGuideRoute>() == true &&
-        !isFavoriteRegionalGuideSelected()
-
-private fun NavBackStackEntry?.isFavoriteRegionalGuideSelected(): Boolean =
-    this != null &&
-        destination.hasRoute<RegionalGuideRoute>() &&
-        toRoute<RegionalGuideRoute>().isFavoriteReentryRoute()
-
-internal fun RegionalGuideRoute.isFavoriteReentryRoute(): Boolean =
-    !initialFavoriteTargetId.isNullOrBlank()
-
-private inline fun <reified T : Any> NavHostController.navigateBottomTabClearingFavoriteReentry(
-    currentBackStackEntry: NavBackStackEntry?,
-    route: T,
-) {
-    if (currentBackStackEntry.isFavoriteRegionalGuideSelected()) {
-        popBackStack<FavoritesRoute>(inclusive = false)
-    }
-
-    navigateBottomTab(route)
-}
-
-private fun NavBackStackEntry?.isItemGuideDetailSource(source: ItemGuideDetailSource): Boolean =
-    this != null &&
-        destination.hasRoute<ItemGuideDetailRoute>() &&
-        toRoute<ItemGuideDetailRoute>().source == source
-
-private fun NavHostController.navigateFavoritesRoot(
-    currentBackStackEntry: NavBackStackEntry?,
-) {
-    when {
-        currentBackStackEntry?.destination?.hasRoute<FavoritesRoute>() == true -> return
-        currentBackStackEntry.isItemGuideDetailSource(ItemGuideDetailSource.FAVORITES) -> {
-            popBackStack<FavoritesRoute>(inclusive = false)
-        }
-        else -> {
-            navigate(FavoritesRoute) {
-                popUpTo(graph.findStartDestination().id) {
-                    saveState = true
-                }
-                launchSingleTop = true
-                restoreState = false
-            }
-        }
-    }
-}
-
-private fun FavoriteCollectionSpotMapMoveRequest.toMapRoute(): MapRoute =
-    MapRoute(
-        favoriteSpotRequestId = UUID.randomUUID().toString(),
-        favoriteSpotTargetId = targetId,
-        favoriteSpotName = name,
-        favoriteSpotType = type.name,
-        favoriteSpotAddress = address,
-        favoriteSpotDetailLocation = detailLocation,
-        favoriteSpotLatitude = latitude,
-        favoriteSpotLongitude = longitude,
-    )
-
-private fun MapRoute.toFavoriteSpotMapMoveRequest(): FavoriteSpotMapMoveRequest? {
-    val targetId = favoriteSpotTargetId ?: return null
-    val name = favoriteSpotName ?: return null
-    val typeName = favoriteSpotType ?: return null
-    val address = favoriteSpotAddress ?: return null
-    val latitude = favoriteSpotLatitude ?: return null
-    val longitude = favoriteSpotLongitude ?: return null
-    val type =
-        runCatching {
-            CollectionSpotType.valueOf(typeName)
-        }.getOrNull() ?: return null
-
-    return FavoriteSpotMapMoveRequest(
-        requestId = favoriteSpotRequestId ?: targetId,
-        targetId = targetId,
-        name = name,
-        type = type,
-        address = address,
-        detailLocation = favoriteSpotDetailLocation,
-        coordinate = Coordinate(
-            latitude = latitude,
-            longitude = longitude,
-        ),
-    )
 }
